web tests working in Docker
